requirements.txt
Wavelink~=2.1.0
python-dotenv~=1.0.0
discord.py~=2.2.3
